requirements.txt
asyncpg==0.23.0
fastapi==0.65.2
tortoise-orm==0.17.4
uvicorn==0.13.4

